+++ types/factories/IGatewayZEVM.sol/IGatewayZEVMEvents__factory.ts
@@ -83,8 +83,8 @@ const _abi = [
           },
           {
             name: "onRevertGasLimit",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
         ],
       },
@@ -189,8 +189,8 @@ const _abi = [
           },
           {
             name: "onRevertGasLimit",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
         ],
       },
@@ -295,8 +295,8 @@ const _abi = [
           },
           {
             name: "onRevertGasLimit",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
         ],
       },
